backend/requirements.txt
boto3
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
django-storages
Django==3.0.8
django-allauth==0.44.0
django-cors-headers==3.3.0
django-crispy-forms==1.10.0
django-extensions==2.2.5
django-heroku==0.3.1
django-model-utils==4.0.0
django-phonenumber-field==5.0.0
django-templated-mail==1.1.1
django-utils-six==2.0
django-widget-tweaks==1.4.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
joblib==1.1.0
numpy==1.21.2
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil
python-decouple==3.5
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==1.0
scipy==1.7.1
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
threadpoolctl==3.0.0
urllib3==1.26.2
whitenoise==5.2.0
